feat: chap2

+++ include/unicodeUtf8.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 uint32_t getByteNumOfEncodeUtf8(int value);
 uint32_t getByteNumOfDecodeUtf8(uint8_t byte);
-uint32_t encodeUtf8(uint8_t *buf, int value);
+uint8_t encodeUtf8(uint8_t *buf, int value);
 int decodeUtf8(const uint8_t *bytePtr, uint32_t length);
 
 #endif
